restructured the folder of component for category route: categoryList.tsx
